Move Settings page component to main Page component
