+++ uclapi-frontend/styles/Settings.module.scss
@@ -1,0 +1,10 @@
+.noPadding {
+  margin-bottom: 0;
+  margin-left: 0;
+  text-align: left;
+}
+
+.title {
+  margin-left: 0;
+  margin: 20px 0;
+}
